Add typecasts to remove compiler warnings.
Former-commit-id: 3895074d3326906e09c16151c853f892809dca18

+++ src/IonHistogram.cpp
@@ -35,7 +35,7 @@ void IonHistogram::addIon(const std::string& name, const double& energy)
         histMap[name] = pTheHist;
     }
     
-    int bin_num = std::floor(energy/binWidth);
+    int bin_num = int(std::floor(energy/binWidth));
     ++(*pTheHist)[bin_num];
 }
 
@@ -74,3 +74,4 @@ void IonHistogram::writeFiles (const std::string& basePath)
         fileStream.close();
     }
 }
+
